[python] endless python troubles

+++ python/notcurses/notcurses-python.h
@@ -196,10 +196,10 @@ static inline void PyObject_cleanup(PyObject **object)
         return_value;                   \
     })
 
-#define GNU_PY_CHECK_UINT(py_function)       \
-    ({                                       \
-        unsigned return_value = py_function; \
-        return_value;                        \
+#define GNU_PY_CHECK_UINT(py_function)                 \
+    ({                                                 \
+        unsigned return_value = (unsigned)py_function; \
+        return_value;                                  \
     })
 
 #define GNU_PY_CHECK_BOOL(py_function)  \
